requirements.txt
Django
djangorestframework
python-dotenv
django-cors-headers
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
whitenoise
daphne
channels
dj_database_url
django-storages[s3]
boto3
requests
celery
social-auth-app-django

# Voice-To-Service pipeline
google-cloud-speech
google-cloud-translate
google-generativeai
mutagen

# Dev / tests
pytest>=8.0
